Reduce tx broadcast interval
Since reconciliation is naturally slower than flooding, and also
since we limit the flooding aspect, we should reduce the intervals
between broadcasts (flooding out or adding to reconciliation sets).
Otherwise, transaction relay will be too slow
(and probably less inefficient).

Note that for privacy reasons the ratio between inbound and outbound
delays matter much more than the actual delays. That ratio is preserved
here, so it is not a privacy degradation.