p2p: Add transactions to reconciliation sets
Transactions eligible for reconciliation are added to the
reconciliation sets. For the remaining txs, low-fanout is used.

If a transaction to be reconciled has in mempool ancestors, we need to be consistent
with the way in which we announce this transaction (either by fanning out or reconciling
all the ancestor set). In order for Erlay to work best, a minimum amount of fanout is required.
Therefore, we cannot simply reconcile with everyone in this case either. Find the subset of peers
that have the least ancestors to be reconciled and send all of them via fanout

Co-authored-by:  Gleb Naumenko <naumenko.gs@gmail.com>